initial commit
I constructed this website as part of my capstone project for the course. It represents the front-end of a Deep Learning project I completed as a component of my MSc dissertation. Currently, it utilizes JavaScript as a temporary placeholder, and I have plans to finalize model deployment and website integration as time allows.